+++ src/kestrel/analytics/manager.py
@@ -1,98 +1,31 @@
-import importlib
-import pkgutil
-import logging
-import inspect
-import sys
-from kestrel.analytics import AbstractAnalyticsInterface
+from kestrel.absinterface import InterfaceManager
+from kestrel.analytics import MODULE_PREFIX, AbstractAnalyticsInterface
 from kestrel.exceptions import (
     AnalyticsInterfaceNotFound,
     InvalidAnalyticsInterfaceImplementation,
     ConflictingAnalyticsInterfaceScheme,
 )
 
-_logger = logging.getLogger(__name__)
 
-
-class AnalyticsManager:
-    def __init__(self, default_schema):
-        self.scheme_to_interface = {}
-        interfaces = _load_analytics_interfaces()
-        for i in interfaces:
-            self.scheme_to_interface.update({s: i for s in i.schemes()})
-
-        self.default_schema = default_schema
-        if default_schema not in self.scheme_to_interface:
-            _logger.error(f"default analytics schema {default_schema} not found.")
-            raise AnalyticsInterfaceNotFound(default_schema)
-
-    def schemes(self):
-        return list(self.scheme_to_interface.keys())
+class AnalyticsManager(InterfaceManager):
+    def __init__(self, config):
+        super().__init__(
+            config,
+            "analytics",
+            ["language", "default_analytics_schema"],
+            MODULE_PREFIX,
+            AbstractAnalyticsInterface,
+            AnalyticsInterfaceNotFound,
+            InvalidAnalyticsInterfaceImplementation,
+            ConflictingAnalyticsInterfaceScheme,
+        )
 
     def list_analytics_from_scheme(self, scheme):
-        scheme = scheme.lower()
-        if scheme not in self.scheme_to_interface:
-            raise AnalyticsInterfaceNotFound(scheme)
-        return self.scheme_to_interface[scheme].list_analytics()
+        i, c = self._get_interface_with_config(scheme)
+        return i.list_analytics(c)
 
     def execute(self, uri, argument_variables, session_id, parameters):
-        scheme, splitter, path = uri.rpartition("://")
-        if not scheme:
-            # If there's only 1 and use didn't specify, use it
-            scheme = self.default_schema
-            if not splitter:
-                uri = self.default_schema + "://" + uri
-            else:
-                uri = self.default_schema + uri
-        scheme = scheme.lower()
-        if scheme not in self.scheme_to_interface:
-            raise AnalyticsInterfaceNotFound(scheme)
-        rs = self.scheme_to_interface[scheme].execute(
-            uri, argument_variables, session_id, parameters
-        )
+        scheme, uri = self._parse_and_complete_uri(uri)
+        i, c = self._get_interface_with_config(scheme)
+        rs = i.execute(uri, argument_variables, c, session_id, parameters)
         return rs
-
-
-def _list_analytics_interfaces():
-    pkg_names = map(lambda x: x.name, pkgutil.iter_modules())
-    itf_names = filter(lambda x: x.startswith("kestrel_analytics_"), pkg_names)
-    return list(itf_names)
-
-
-def _load_analytics_interfaces():
-    interface_names = _list_analytics_interfaces()
-    interfaces = []
-    for interface_name in interface_names:
-        importlib.import_module(interface_name)
-        cls = inspect.getmembers(sys.modules[interface_name], _is_interface_class)
-        if not cls:
-            raise InvalidAnalyticsInterfaceImplementation(
-                f'no interface class found in "{interface_name}"'
-            )
-        elif len(cls) > 1:
-            raise InvalidAnalyticsInterfaceImplementation(
-                f'more than one interface class found in "{interface_name}"'
-            )
-        else:
-            interface = cls[0][1]
-            interface_conflict, scheme_conflict = _find_scheme_conflict(
-                interface, interfaces
-            )
-            if interface_conflict:
-                raise ConflictingAnalyticsInterfaceScheme(
-                    interface, interface_conflict, scheme_conflict
-                )
-            interfaces.append(interface)
-    return interfaces
-
-
-def _is_interface_class(obj):
-    return inspect.isclass(obj) and obj.__bases__[0] == AbstractAnalyticsInterface
-
-
-def _find_scheme_conflict(new_interface, interfaces):
-    for interface in interfaces:
-        for scheme_new in new_interface.schemes():
-            for scheme_old in interface.schemes():
-                if scheme_new == scheme_old:
-                    return interface, scheme_new
-    return None, None