+++ CHANGELOG.rst
@@ -6,6 +6,59 @@ All notable changes to this project will be documented in this file.
 
 The format is based on `Keep a Changelog`_.
 
+1.2.0 (2022-02-10)
+==================
+
+Added
+-----
+
+- Kestrel main package
+
+  - matplotlib figure support in Kestrel Display Objects
+  - analytics interface upgraded with config shared to Kestrel
+    
+- Python analytics interface
+
+  - minimal requirement design for writing a Python analytics
+  - analytics function environment setup and destroy
+  - support for a variety of display object outputs
+  - parameters support
+  - stack tracing for exception inside a Python analytics
+    
+- STIX-shifter data source interface
+
+  - automatic STIX-shifter connector install
+    
+    - connector name guess
+    - connector origin verification
+    - comprehensive error and suggestion if automatic install failed
+        
+  - pretty print for exception inside a Docker analytics
+    
+- documentation
+
+  - Python analytics interface
+  - Kestrel debug page
+  - flag to disable certificate verification in STIX-shifter profile example
+
+Changed
+-------
+
+- abstract interface manager between datasource/analytics for code reuse
+
+Fixed
+-----
+
+- auto-complete with data source #163
+- exception for empty STIX-shifter profile
+- STIX-shifter profile name should be case insensitive
+- exception inappropriately caught when dereferencing vars with no time range
+
+Removed
+-------
+
+- documentation about STIX-shifter connector install
+
 1.1.7 (2022-01-27)
 ==================